UI sketches and Class Diagram
This is the first draft of the UI and the class diagram. I abstracted the room chart from the room, I'm not sure if that's the right thing to do or not. Essentially it starts with the room chart list that mirrors what the House screen does in the PMS. Double clicking on one of those rows will bring up the Room detail (page 2 of the UI sketch). From there the user can click begin cleaning and finish cleaning.