SimpleC#ConsoleApp
This is just a simple console app I made with C# just to get a feel for the language and also to flush out some of the basic logic of what this program will become. Essentially this operates as the screen after the user drills into a room ( see page 2 of UI sketch).