Renamed all variables with proper conventions and Implemented TimeIn, Timeout, and Houskeeper Selection
TimeIn and TimeOut update the database with a DateTime Object, however notice in SQLCommands.cs --> Line147 I used the GETDATE() sql call to update the database. The method is not taking in a DateTime object from the rest of the program. When I tried to do that I was getting a syntax error for the C# SQL library. I'm sure there's a better way to do this. Also the HskprName field is in the database and is properly updated from the user's initial selection in the program.