ElapsedTime field is now updated in the database
I had to pull the TimeIn value from the database using a method called SQLGetElapsedTime due to the way the UI works. A new room object is created after choosing:
 1. Begin Cleaning Room
2. Finish Cleaning Room.
Therefore, the TimeIn class member is  never strored in the same object instance as the TimeOut class member. I don't think it's that big of a deal for this test app but for the real thing it will need to be corrected.