RoomGroup Datastructure implementation
I created 3 classes here. FirstFloor and SecondFloor, which initilizes a list of rooms with a GroupID string (i.e. "1st Floor"), and a RoomGroup Class who's only class member is a Dictionary object that maps the GroupID string to the a list rooms. The RoomGroup Class initializes the FirstFloor and SecondFloor classes, thus creating the list of rooms based upon which floor they're on and their respective GroupIDs.