Added the dictionary for the Initilization of the RoomGroups
The RoomGroup.cs class contains two class members: GroupID (string) and RmNmbrs (List<string>). This class retrieves a list of rooms given a GroupID(i.e. "1st Floor" will retrieve all room numbers "101" - "112" as a string. The GroupID and RmNmbrs objects are then set to the class members so they may be referenced by the RoomGroup.cs class. This class contains one class member: Dictionary<string, List<string>>. This is the globally accessible dictionary to that given a key (which is the GroupID) it can retrieve a List<string> of RmNmbrs. The main program then initializes each floor.